tech-debt: split demo-test.controller into per-domain sub-controllers (ROK-1071)
Split the 358-line DemoTestController into 5 domain-scoped sub-controllers
so each stays well under the 300-line cap:

- demo-test-voice.controller.ts (56) — voice sessions, classify
- demo-test-scheduled-events.controller.ts (73) — enable/disable/cleanup/trigger-completion, pause-reconciliation, set-event-times
- demo-test-signups.controller.ts (73) — signup, trigger-departure, cancel-signup
- demo-test-games.controller.ts (125) — add/clear game interest, steam-app-id, get-game, auto-heart, steam-nudge, cancel-lineup-phase-jobs
- demo-test-core.controller.ts (117) — link-discord, enable-notifications, notifications, flush-notification-buffer, flush-embed-queue, await-processing, clear-game-time-confirmation

All routes remain under /admin/test/* with AuthGuard('jwt') + AdminGuard.
Extracted the shared parseBody helper into demo-test.utils.ts. Tests split
alongside each controller. Service is unchanged (service split is ROK-1072).
